Remove today's date from individual entries
We assume we're looking at room bookings for today only, so remove the redundant date from each entry, and instead add today's date to the top of the page. Seeing today's date at the top of the pagre helps the user know the calendar display is up to date and not stale.