Refresh the dashboard every 5 minutes
We want the dashboard to be an up to date view of the room bookings.
Using a low-fi meta tag to refresh the page every 5 minutes is a decent
enough solution for now. If we have trouble with intermittent network
connections or API availability we could use a more complex approach
where we don't lose the old data on refresh.