Filter out events the rooms have declined
Previously those events were shown in the same way as accepted events,
giving the impression of clashes and of reservations where there are
none.

An alternative solution would be to present declined events differently,
but I cannot see the value that seeing those declined events brings,
given the use of the dashboard.