Handle events with no attendees
Added a guard clause to prevent nils causing errors.

There was a private event in one of the rooms that had no attendees
(returned by the calendar API as `nil` rather than an empty array). This
caused `.find(&:self)` to be called on nil which caused the app to blow
up.