Improve displayed summary of private events
Private events have a nil `summary`.

Previously, the app was showing `’Unspecified’` when it encountered a nil summary.

This text confused some colleagues, who thought it meant the room was unused. This created awkwardness and upset for both the person who had booked the room and the person “squatting” in the room.

This commit changes the fallback text to `’Private or unspecified’` for nil summary and organiser fields.

Hopefully this will make it less confusing when a private event is displayed.