Test environment setup
The meeting room dashboard started as a tiny Sinatra app with very small
functionality, and no tests. It now does a lot more, and with added complexity
comes a higher risk of something breaking, but it still doesn't have
automated tests.

This commit introduces the initial configuration for testing the app, both API
and user-facing app. We will be using Rspec, Rack-test, WebMock and Capybara,
plus some funcionality from Activesupport for time-dependant tests.

Tests will also be run in the CI build on every push and PR to master branch.