Rspec and Rack-test setup
The meeting room dashboard started as a very tiny Sinatra repo with very small
functionality, and no tests. It now does a lot more, and with added complexity
comes a higher risk of something breaking, but it still doesn't have
automated tests.

This commit introduces the initial configuration for testing with Rspec and
Rack-Test.