Tests for the API
This commit introduces a test suite for the API side of the app. I created a
Helpers module to include the methods used in the API tests as I wanted them to
be available for the user-facing app tests too.

These tests showed an error in the code when it comes to reject meetings with
no attendees or when all attendees have declined. I fixed the error in
`lib/room.rb` and now the tests pass.