Tests for user-facing app
This commit introduces a test suite for the user-facing app (dashboard),
for both a single room view and a board (multi-room) view.
Some tests are time-dependant so a new method has been added to the Helpers
module to allow me to test for a scenario with current ongoing meetings.

Please note I added a value for `datetime_time` on `multi_room.haml` for the
purpose of testing the `show_clock` funcionality. This functionality is given
by Javascript code. I tried to test the Javascript, using three different drivers
and a few different gems, but the clock test kept failing. I believe this could
be due to the js function being a permanent loop that never ends (it refreshes
the clock every ten seconds). I paired with a Javascript-experienced
senior developer for this test and after trying unsuccessfully to make a test
work, we decided that since this is a small piece of functionality that does
not compromise the security and stability of the app, we could afford not to
test the JS, given the difficulty and amount of resources needed for this
single test. I did a test for the id tag to be present and the value passed by
Ruby (the interpolation on the haml file) to be present. This will allow us to
have at least a bit of test coverage for `show_clock` functionality.

I have also removed a JS line at the bottom that calls a method that has not
been defined, with a value that is not present anywhere in the code and that
makes no difference to the behaviour of the app. This is probably something
someone forgot to delete in the past.