Room and multi_room views refactored
Since part of the code on both room and multi_room haml view files is the same,
I decided to refactor this and include that code in a partial that will be
rendered on those views. This avoids having duplicated code, but also prevents
from having to implement unnecessary/duplicated tests for the dashboard if
changes to these files are introduced in the future.