(Fix) Fix test which fails on single-digit days
This test is comparing a date output to an HTML page with a date in the test.
When the day is single digit, `Date.today.strftime("%A, %B %e")` will pad
the day with a space, e.g. `Tuesday, January  5`; but in the HTML output
the test is comparing, the extra space is stripped out by the browser.

Use `Date.today.strftime("%A, %B %-d")` instead which won't pad a single digit.